refactor: 검색 컴포넌트 이동
